Ombra v2
Ombra is a self-driving Arduino car that has a clap-on clap-off switch. A sonar sensor is attached to the front of the Arduino car so that when obstacles are detected, the car goes in reverse mode and steers to the right. There is also a RGB LED strip attached to the bottom of the car which will indicate green when the car is going forward and red when in reverse. The final working prototype is 3d printed and painted with automotive paint.